+++ packages/form-data-middleware/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`form-data-middleware`](https://github.com/remix-run/remix/tree/main/packages/form-data-middleware). It follows [semantic versioning](https://semver.org/).
 
+## Unreleased
+
+- Explicitly set `context.formData` in all `POST` cases, even when the request body is invalid
+
 ## v0.1.0 (2025-11-19)
 
 Initial release extracted from `@remix-run/fetch-router` v0.9.0.